functions/convo_live_mark_word_mastered/requirements.txt
functions-framework==3.*
google-cloud-firestore==2.*
flask==2.*
